requirements.txt
Flask==2.3.3
Werkzeug==2.3.8
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.3.0
blinker==1.9.0
gunicorn==21.2.0
Flask-Mail==0.9.1
psycopg2-binary==2.9.7
